Change the infinite sites wording
While discussing https://github.com/tskit-dev/msprime/issues/1600 I noticed that the text describing infinite sites could be a bit confusing, as it refers to the "infinite sites model", which is not one of the "mutation model"s that we actually list. So perhaps we should refer to it as the "infinite sites assumption" instead (and indeed, this makes it a little more obvious that any of the mutation models we have can be combined with an IS assumption?

I've also put it in a `{note}` for greater emphasis, as I think it's pretty important. But we could even put it in an `{important}` block instead?